Add Expanded Option of Smart Primary Selection
With gradual code changes with adding armor values, it seems that primary selection was never updated to take those into account (nor did it include checks to the subsystem). This makes AI not really choose the best weapon for a situation if armor tables are used heavily to determine balance. This PR adds an updated logic in `ai_select_primary_weapon` along with an `ai_profiles` flag to turn it on. I'd also be open to allowing the threshold values used in the function as settable for the modder in `ai_profiles`.