Add multi support to external SEXPs
This adds (mostly) automatic multi support to the external SEXP system.
This will record the parameter values that were evaluated for the master
and automatically send them to the client via the standard SEXP packet
system.

For SEXPs where the behavior does not differ between master and client
there needs to be no special handling of the multi case in the SEXP
action code.